Move LustrumController to Minisite and fix controller
